Patter.jl

Patter.jl

Functions

Index

Functions

Patter.assemble_yobs — Method

# Assemble `yobs` for the particle filter

Assemble a dictionary of observations (and associated model parameters) for the particle filter.

Arguments

datasets: A Vector of DataFrames, one for each data type. Each DataFrame must contain the following columns:

timestamp: A DateTime Vector of time stamps;

sensor_id: A vector of sensor IDs;

obs: The observation;

Additional columns required to construct ModelObs instances (that is, model parameters);

model_types: A Vector of ModelObs subtypes for each dataset.

Details

The function iterates over datasets and models and creates a typed dictionary of timestamps. Each time step contains a vector of Tuples, with one element for each sensor that recorded an observation at that time stamp. Each element is a tuple that defines the observation and the model parameters (that is, a ModelObs instance).

Patter.distance — Function

distance(x0::Real, y0::Real, x1::Real, y1::Real)

Calculate Euclidean distances between coordinates.

Arguments

x0, y0: The coordinates of the first point;

x1, y1: The coordinates of the second point;

Returns

A number that defines the distance between two coordinates.

Patter.extract — Method

extract(map::GeoArrays.GeoArray, x::Real, y::Real)

Extract the value of a GeoArray (such as a bathymetry grid) at a specififed pair of x and y coordinates.

Details

In Patter.jl, NaN elements are taken to define inhospitable habitats, such as land (see is_valid()).

To extract a value from a map in another format, such as a shapefile, write a custom extract() method.

Returns

The value of the GeoArray for a coordinate pair within the bounds of map;

NaN (of the same type as map's elements) for a coordinate pair beyond the bounds of map;

Patter.is_valid — Function

is_valid(map_value::Real)
is_valid(map_valu::Real, z::Real)

Determine the validity of a point on a map.

For 2D (x, y) states, is_valid(map_value) checks if the map_value at point (x, y) is not NaN.

For 3D (x, y, z) states, is_valid(map_value, z), checks that map_value at point (x, y) is not NaN, and also ensures that the provided z-coordinate (z) lies within the valid range, specifically 0 < z ≤ map_value.

Arguments

map_value: The map from which z-coordinates are extracted;

x: The x-coordinate of the point;

y: The y-coordinate of the point;

z: The z-coordinate of the point to be validated (only for the 3D case);

Returns

true if the conditions for validity are met;

false otherwise;

Patter.julia_get — Function

Julia from R

A collection of functions that facilitate the translation of inputs from R into Julia.

Details

[julia_get_xinit()] gets a Vector of initial States from a DataFrame;

[julia_get_model_types()] gets a Vector of ModelObs subtypes from a Vector of Strings;

[julia_get_models()] gets a Vector of model instances from a Vector of DataFrames that contain parameters and a corresponding vector of model types;

Patter.julia_get_xinit — Function

Julia from R

A collection of functions that facilitate the translation of inputs from R into Julia.

Details

[julia_get_xinit()] gets a Vector of initial States from a DataFrame;

[julia_get_model_types()] gets a Vector of ModelObs subtypes from a Vector of Strings;

[julia_get_models()] gets a Vector of model instances from a Vector of DataFrames that contain parameters and a corresponding vector of model types;

Patter.logpdf_obs — Function

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.particle_filter — Method

Particle filter

A particle filtering algorithm that samples from f(Xt | {Y1 ... Y_t}) for t ∈ 1:Tmax.

Arguments:

timeline: A Vector{DateTime} of time stamps that defines the time steps for the simulation. Time stamps must be equally spaced.

xinit: A Vector{State} that defines the initial state(s) of the animal.

yobs: A Dictionary of observations.

move: A ModelMove instance.

n_move: An integer that defines the number of attempts used to find a legal move. Particles are killed otherwise.

n_record: An integer that defines the number of particles to record at each time step.

n_resample: A number that defines the effective sample size at which to resample particles.

Returns

A tuple with the following fields:

timeline

state

ess

maxlp

convergence

Patter.r_get_states — Method

R from Julia

A collection of functions that facilitate the translation of inputs from Julia into R.

Details

[r_get_states] translates a State matrix into a DataFrame that can be passed to R. In the input matrix, each row is a particle and each column is a time step.

[r_get_dataset] translates a Dictionary of observations into a Vector of DataFrames that can be passed to R.

Returns

A long-format DataFrame, with columns for path_id, timestep and each state dimension.

Patter.resample — Function

Systematic resampling algorithm

Given the weight vector w, resample a set of indices based on low-variance resampling algorithm from Thrun, Burgard, and Fox's "Probabilistic Robotics".

Source

Code adapted from https://github.com/JuliaStats/StatsBase.jl/issues/124.

Example

X = ["A", "B", "C", "D"]
w = [0, 0, 0.75, 0.25]

idx = resample(w, 12)
X[idx]

Patter.simulate_move — Function

simulate_move(state::State, model::MoveModel, t::Int64, n_trial::Real)

Simulate movement from one location (State) into a new location (State).

Details

simulate_move() is an internal function that uses an Patter.simulate_step() method to simulate proposals for a new state until a valid proposal is generated;

Patter.simulate_obs — Function

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.simulate_path_walk — Method

Simulate movement paths

This function simulates discrete-time movement path(s) from a vector of initial states and random-walk movement model.

Arguments

xinit: A Vector of States that defines the initial state(s) for the simulation;

move: A ModelMove instance;

nt: An integer that defines the number of time steps;

Details

For each initial state, a movement path is simulated.

Returns

A matrix:

Rows: paths

Columns: time steps

Patter.simulate_state_init — Function

simulatestateinit(state::State, model::MoveModel, xlim, ylim)

Simulate an initial state.

This function is wrapped by the exported function simulate_states_init(), which simulates a vector of states.

Arguments:

state_type: An empty State subtype, such as StateXY, used for method dispatch only;

model: A MoveModel instance;

xlim, ylim: Pairs of numbers that define the boundaries of the area within which x and y state values are sampled;

Patter.simulate_states_init — Function

Simulate initial states

Simulate a vector of initial states for the simulation of movement paths and the particle filter.

Arguments

state_type: An empty State subtype, such as StateXY, used for method dispatch only;

move: A MoveModel instance;

n: The number of intial states to simulate;

xlim, ylim: (optional) Pairs of numbers that define the boundaries of the area within which x and y state values are sampled;

Patter.simulate_step — Function

simulate_step(state::State, model::ModelMove, t::Int64)

Simulate a (tentative) step from one location (State) into a new location (State).

Details

simulate_step() is a generic function that simulates a new value for the animal's state;

Different methods are dispatched according to the state and the movement model;

New methods must be provided for custom states or movement models;

Internally, simulate_step() is wrapped by simulate_move(), which implements simulate_step() iteratively until a valid proposal is generated;

Patter.simulate_yobs — Method

Simulate observations

For each simulated path, simulate a dictionary of observations.

Arguments

paths: A Matrix of simulated paths; from simulate_path_walk();

models: A Vector of ModelObs instances;

timeline: A Vector{DateTime} of time stamps;

Returns

A dictionary, with one entry for each path;

Each entry is a dictionary, with one entry for each time stamp;

Each time stamp entry is a Vector of Tuples, each comprising the observation and the associated ModelObs instance;

Patter.state_is_valid — Method

state_is_valid(state::State, zdim::Bool)

Determine whether or not a state is valid.

See also State, is_valid(),

Patter.ModelMove — Type

Movement models

ModelMove is an abstract type that defines the movement model.

Subtypes

MoveModelXY: A subtype for two-dimensional (x, y) random walks, based distributions for step lengths (dbn_length) and turning angles (dbn_angle);

MoveModelXYZ: A subtype for three-dimensional (x, y, z) random walks, based on distributions for step lengths (dbn_length), turning angles (dbn_angle) and changes in depth (dbn_z_delta);

MoveModelXYZD: A subtype for four-dimensional (correlated) random walks, based on distributions for step lengths (dbn_length), changes in turning angle (dbn_angle) and changes in depth (dbn_z_delta);

Fields

map: A field that defines the arena within which movement occurs. This is required by all movement models;

dbn_length: The distribution of step lengths;

dbn_angle: The distribution of turning angles;

dbn_angle_delta: The distribution of changes in turning angle;

dbn_z_delta: The distribution of changes in depth;

Details

ModelMove structures define the parameters of the movement model;

All ModelMove structures must contain an map field.

By default, map is assumed to be a GeoArray but a shapefile can be used with a custom extract() method;

Users can use a provided structure or write their own;

For custom ModelMoves, new rstep() methods are required;

Patter.ModelMoveXY — Type

Movement models

ModelMove is an abstract type that defines the movement model.

Subtypes

MoveModelXY: A subtype for two-dimensional (x, y) random walks, based distributions for step lengths (dbn_length) and turning angles (dbn_angle);

MoveModelXYZ: A subtype for three-dimensional (x, y, z) random walks, based on distributions for step lengths (dbn_length), turning angles (dbn_angle) and changes in depth (dbn_z_delta);

MoveModelXYZD: A subtype for four-dimensional (correlated) random walks, based on distributions for step lengths (dbn_length), changes in turning angle (dbn_angle) and changes in depth (dbn_z_delta);

Fields

map: A field that defines the arena within which movement occurs. This is required by all movement models;

dbn_length: The distribution of step lengths;

dbn_angle: The distribution of turning angles;

dbn_angle_delta: The distribution of changes in turning angle;

dbn_z_delta: The distribution of changes in depth;

Details

ModelMove structures define the parameters of the movement model;

All ModelMove structures must contain an map field.

By default, map is assumed to be a GeoArray but a shapefile can be used with a custom extract() method;

Users can use a provided structure or write their own;

For custom ModelMoves, new rstep() methods are required;

Patter.ModelMoveXYZD — Type

Movement models

ModelMove is an abstract type that defines the movement model.

Subtypes

MoveModelXY: A subtype for two-dimensional (x, y) random walks, based distributions for step lengths (dbn_length) and turning angles (dbn_angle);

MoveModelXYZ: A subtype for three-dimensional (x, y, z) random walks, based on distributions for step lengths (dbn_length), turning angles (dbn_angle) and changes in depth (dbn_z_delta);

MoveModelXYZD: A subtype for four-dimensional (correlated) random walks, based on distributions for step lengths (dbn_length), changes in turning angle (dbn_angle) and changes in depth (dbn_z_delta);

Fields

map: A field that defines the arena within which movement occurs. This is required by all movement models;

dbn_length: The distribution of step lengths;

dbn_angle: The distribution of turning angles;

dbn_angle_delta: The distribution of changes in turning angle;

dbn_z_delta: The distribution of changes in depth;

Details

ModelMove structures define the parameters of the movement model;

All ModelMove structures must contain an map field.

By default, map is assumed to be a GeoArray but a shapefile can be used with a custom extract() method;

Users can use a provided structure or write their own;

For custom ModelMoves, new rstep() methods are required;

Patter.ModelObs — Type

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.ModelObsAcousticLogisTrunc — Type

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.ModelObsDepthNormalTrunc — Type

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.ModelObsDepthUniform — Type

Observation models

Structures

ModelObs is an abstract type used to hold parameters for observation models. For communication with R, all subtypes should include a sensor_id field.

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation and a truncated logistic detection probability model:

sensor_id: The receiver ID;

receiver_x, receiver_y: The coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: The parameters of a logistic detection probability model;

ModelObsDepthUniform is ModelObs structure for a depth observation and a uniform depth model, in which we assume the individual must be located in an envelope around the bathymetry depth, defined by two error terms:

sensor_id: The sensor ID;

depth_shallow_eps

depth_deep_eps

ModelObsDepthNormalTrunc is a ModelObs structure for a depth observation and a truncated normal model, in which we assume the likelihood of a depth observation given by a normal distribution centred at the bathymetric depth in a location and defined by the parameters:

sensor_id: The sensor ID;

depth_sigma: The standard deviation of the normal distribution;

deep_depth_eps: The deep truncation parameter;

Simulation

simulate_obs() is a generic function that simulates observations from a ModelObs instance.

Density

logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's state and a ModelObs instance.

Patter.State — Type

State

State is an abstract type that defines the animal's state at a given time step.

Subtypes

StateXY: Used for two dimensional (x, y) states;

StateXYZ: Used for three-dimensional (x, y, z) states;

StateXYZD: Used for four-dimensional (x, y, z, direction) states;

Fields

map_value: The value of the map at coordinates (x, y), required for All States;

x, y: The animal's x and y coordinates, required for all States;

z: The animal's z coordinate, required for 3D States;

angle: The turning angle, required by StateXYZD;

Details

All states must include x, y and map_value fields;

For >= 3D states, the depth dimension must be named z (for Patter.simulate_move());

For R users, all fields must be of type Float64 for Patter.r_get_states() to parse state vectors;

```

Patter.StateXY — Type

State

State is an abstract type that defines the animal's state at a given time step.

Subtypes

StateXY: Used for two dimensional (x, y) states;

StateXYZ: Used for three-dimensional (x, y, z) states;

StateXYZD: Used for four-dimensional (x, y, z, direction) states;

Fields

map_value: The value of the map at coordinates (x, y), required for All States;

x, y: The animal's x and y coordinates, required for all States;

z: The animal's z coordinate, required for 3D States;

angle: The turning angle, required by StateXYZD;

Details

All states must include x, y and map_value fields;

For >= 3D states, the depth dimension must be named z (for Patter.simulate_move());

For R users, all fields must be of type Float64 for Patter.r_get_states() to parse state vectors;

```

Patter.StateXYZ — Type

State

State is an abstract type that defines the animal's state at a given time step.

Subtypes

StateXY: Used for two dimensional (x, y) states;

StateXYZ: Used for three-dimensional (x, y, z) states;

StateXYZD: Used for four-dimensional (x, y, z, direction) states;

Fields

map_value: The value of the map at coordinates (x, y), required for All States;

x, y: The animal's x and y coordinates, required for all States;

z: The animal's z coordinate, required for 3D States;

angle: The turning angle, required by StateXYZD;

Details

All states must include x, y and map_value fields;

For >= 3D states, the depth dimension must be named z (for Patter.simulate_move());

For R users, all fields must be of type Float64 for Patter.r_get_states() to parse state vectors;

```

Patter.StateXYZD — Type

State

State is an abstract type that defines the animal's state at a given time step.

Subtypes

StateXY: Used for two dimensional (x, y) states;

StateXYZ: Used for three-dimensional (x, y, z) states;

StateXYZD: Used for four-dimensional (x, y, z, direction) states;

Fields

map_value: The value of the map at coordinates (x, y), required for All States;

x, y: The animal's x and y coordinates, required for all States;

z: The animal's z coordinate, required for 3D States;

angle: The turning angle, required by StateXYZD;

Details

All states must include x, y and map_value fields;

For >= 3D states, the depth dimension must be named z (for Patter.simulate_move());

For R users, all fields must be of type Float64 for Patter.r_get_states() to parse state vectors;

```

Index

Patter.ModelMove

Patter.ModelMoveXY

Patter.ModelMoveXYZD

Patter.ModelObs

Patter.ModelObsAcousticLogisTrunc

Patter.ModelObsDepthNormalTrunc

Patter.ModelObsDepthUniform

Patter.State

Patter.StateXY

Patter.StateXYZ

Patter.StateXYZD

Patter.assemble_yobs

Patter.distance

Patter.extract

Patter.is_valid

Patter.julia_get

Patter.julia_get_xinit

Patter.logpdf_obs

Patter.particle_filter

Patter.r_get_states

Patter.resample

Patter.simulate_move

Patter.simulate_obs

Patter.simulate_path_walk

Patter.simulate_state_init

Patter.simulate_states_init

Patter.simulate_step

Patter.simulate_yobs

Patter.state_is_valid