Patter.jl

Patter.jl

Functions

Index

Functions

Patter.assemble_yobs — Method

assemble_yobs(; datasets::Vector, model_obs_types::Vector{DataType})

Assemble a dictionary of observations (and associated model parameters) for the particle filter (particle_filter()).

Arguments (keywords)

datasets: A Vector of DataFrames, one for each data type. Each DataFrame must contain the following columns:

timestamp: A DateTime Vector of time stamps;

sensor_id: A Vector of sensor IDs;

obs: A Vector of observations;

Additional columns required to construct ModelObs instances (that is, model parameters);

model_obs_types: A Vector of ModelObs sub-types (one for each dataset);

Details

The function iterates over animal-tracking datasets (for example, acoustic and archival [depth] time series for a particular individual) and corresponding observation model types and creates a typed dictionary of time stamps for the particle filter (see particle_filter()'s yobs argument). Each time step contains a Vector of Tuples, with one element for each sensor that recorded an observation at that time stamp. For example, each element might correspond to an acoustic receiver and/or the individual's archival tag. Each element is a Tuple that defines the observation and the corresponding observation model parameters (that is, a ModelObs instance).

Returns

A Dict:

Each element corresponds to a timestamp:

Each timestamped element contains a Vector of Tuples (one for each observation):

Each Tuple contains an observation and the corresponding ModelObs instance;

See also

assemble_yobs() to assemble real-world datasets for the particle filter;

simulate_yobs() to simulate observations for the particle filter;

particle_filter() to implement the particle filter;

Patter.extract — Method

extract(map::GeoArrays.GeoArray, x::Real, y::Real)

(Internal) Extract the value of a GeoArray (such as a bathymetry grid) at a pair of x and y coordinates.

Details

In Patter.jl, map is a GeoArray that defines the area within which movements are possible. In our applications, map is often a bathymetry raster that defines the depth of the seabed across the study area. map is an essential field in individual movement models (see ModelMove). The internal extract() function supports the simulation of initial states (via simulate_states_init()) and the updating of states (via Patter.simulate_step()), as required to simulate animal movement paths (e.g., in simulate_path_walk() and particle_filter()). Individual states define, at a minimum, the individual's location (x, y) and the value of the map at that location (map_value), which is extracted by extract() (see State). The Coordinate Reference Systems for map, x and y must align for this to work (i.e., map should use a Universal Transverse Mercator projection with coordinates in metres). extract() is exported so that it can be used in new methods (for custom states or movement models) of these functions. The simulation of individual movements (via Patter.simulate_step()) is implemented iteratively (via Patter.simulate_move()) until a valid movement is found. NaN is taken to define inhospitable habitats, such as land, into which the individual cannot move (see is_valid()). It should be possible to use a map in another format (such as a shapefile) within these routines, with a custom extract() method that returns NaN in inhospitable habitats and a numeric constant otherwise.

Returns

The value of the GeoArray for a coordinate pair within the bounds of map;

NaN (of the same type as map's elements) for a coordinate pair beyond the bounds of map;

See also

State for State sub-types;

Patter.extract() to extract values from map at State coordinates;

Patter.simulate_step() to simulate a new State;

Patter.is_valid() to determine whether or not a simulated state is valid;

Patter.simulate_move() to simulate states iteratively until a valid state is found;

simulate_path_walk() and particle_filter() for the front-end functions that use these routines to simulate animal movement paths;

Patter.is_valid — Function

is_valid(map_value::Real)
is_valid(map_value::Real, z::Real)

(Internal) Determine the validity of a point on a map.

For two-dimensional (x, y) states, is_valid(map_value) checks if the map_value (at a point (x, y)) is not NaN.

For states with a depth (z) component, is_valid(map_value, z), checks that map_value is not NaN and that the provided z-coordinate (z) lies within the valid range, specifically 0 < z ≤ map_value (i.e., the animal is not below the seabed).

Arguments

map_value: The value of the map in a particular location;

z: (optional) The z-coordinate (depth) of the animal, used for states that contain a depth component only;

Details

These are internal functions. They are used to validate simulated individual states (see State). Individual states (i.e., locations) are simulated via Patter.simulate_step(). Patter.simulate_move() wraps Patter.simulate_step(), iteratively proposing states until a valid state is found. This is required to simulate animal movement paths (e.g., in simulate_path_walk() and particle_filter()).

Returns

true if the conditions for validity are met;

false otherwise;

See also

State for State sub-types;

Patter.extract() to extract values from map at State coordinates;

Patter.simulate_step() to simulate a new State;

Patter.is_valid() to determine whether or not a simulated state is valid;

Patter.simulate_move() to simulate states iteratively until a valid state is found;

simulate_path_walk() and particle_filter() for the front-end functions that use these routines to simulate animal movement paths;

Patter.logpdf_move — Method

logpdf_move(state_from::State, state_to::State, state_zdim::Bool, 
            model_move::ModelMove, 
            t::Int, 
            vmap::Union{GeoArray, Nothing}, 
            n_sim::Int,
            cache::Union{Dict, Nothing})

(Internal) Evaluate the log probability of a movement between two States (state_from and state_to).

Arguments

state_from: A State instance that defines a State from which the animal moved;

state_to: A State instance that defines a State into which the animal moved;

state_zdim: A Boolian that defines whether or not state_from and state_to contain a z (depth) dimension;

model_move: A ModelMove instance;

t: An integer that defines the time step;

vmap: (optional) A GeoArray that maps the region within which movements from state_from are always legal. Valid regions must equal 1. vmap can be provided for 'horizontal' movement models (e.g., if state_from and state_to are StateXY instances (see State));

n_sim: An integer that defines the number of Monte Carlo simulations (used to approximate the normalisation constant);

cache: (optional) A Dict of normalisation constants including state_from;

Details

Patter.logpdf_move() is an internal function that evaluates the log probability of a movement step between two State(s) (i.e., locations). This function wraps Patter.logpdf_step(), accounting for accounting for restrictions to movement; that is, Patter.logpdf_move() evaluates logpdf_step(state_from, state_to, model_move, t, length, heading) + log(abs(determinate)) - log(Z) where Z is the normalisation constant. If model_move is 'horizontal (e.g., state_from and state_to are two-dimensional, StateXY instances), a 'validity map' (vmap) can be provided. This is a GeoArray that define the regions within which movements between two locations are always legal. In the case of an aquatic animal, this is the region of the study area that is the sea, shrunk by state_from.mobility. In this instance, the normalisation constant is simply log(1.0). Otherwise, a Monte Carlo simulation of n_sim iterations is required to approximate the normalisation constant, accounting for invalid movements, which is more expensive (see logpdf_move_normalisation()). Set n_sim = 0 and cache = nothing for unrestricted models (i.e., if particle_filter() was implemented with n_move = 1). Patter.logpdf_move() is used for particle smoothing (see particle_smoother_two_filter()).

Returns

A number (log probability);

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() and Patter.simulate_move() to simulate new States;

Patter.logpdf_step() and Patter.logpdf_move() to evaluate the log-probability of movement between two locations;

Patter.logpdf_move_normalisation() for estimation of the normalisation constant;

particle_smoother_two_filter() for the front-end function that uses these routines for particle smoothing;

Patter.logpdf_move_normalisation — Method

logpdf_move_normalisation(state::State, state_zdim::Bool, 
                          model_move::ModelMove, t::Int, 
                          vmap::Union{GeoArray, Nothing},
                          n_sim::Int)

(Internal) Approximate the (log) normalisation constant for the (log) probability density of movement from one State (location) into another.

Arguments

state: A State instance that defines a State from which the animal moved;

state_zdim: A Boolian that defines whether or not state contains a z (depth) dimension;

model_move: A ModelMove instance;

t: An integer that defines the time step;

vmap: (optional) A GeoArray that maps the region within which movements from state are always legal. Valid regions must equal 1. vmap can be provided for 'horizontal' movement models (e.g., if state is a StateXY);

n_sim: An integer that defines the number of Monte Carlo simulations. Set n_sim = 0 if n_move = 1 in particle_filter();

Details

This internal function computes the normalisation constant for the (log) probability of movement from one State (state) into another. If n_move = 1 in particle_filter(), set n_sim = 0 to return log(1.0). Otherwise, we need to account for the truncation of the movement model by land. If model_move is 'horizontal' (e.g., state is a two-dimensional, StateXY instance), a 'validity map' (vmap) can be provided. This is a GeoArray that define the regions within which movements from that state are always legal. In the case of an aquatic animal, this is the region of the study area that is the sea, shrunk by state.mobility. In this instance, the normalisation constant is simply log(1.0). Otherwise, a Monte Carlo simulation of n_sim iterations is used to estimate the normalisation constant. A Beta(1, 1) prior is used to correct for simulations that fail to generate valid move from state. This function is used by Patter.logpdf_move() to evaluate the (log) probability of movement between two states, which is required for particle smoothing (see particle_smoother_two_filter()).

Returns

A number (the log normalisation constant);

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() and Patter.simulate_move() to simulate new States;

Patter.logpdf_step() and Patter.logpdf_move() to evaluate the log-probability of movement between two locations;

Patter.logpdf_move_normalisation() for estimation of the normalisation constant;

particle_smoother_two_filter() for the front-end function that uses these routines for particle smoothing;

Patter.logpdf_step — Function

logpdf_step(state_from::State, state_to::State, model_move::ModelMove, length, heading)

(Internal) Evaluate the (unnormalised) log probability of an (unrestricted) movement step.

Arguments

state_from: A State instance that defines a State from which the animal moved;

state_to: A State instance that defines a State into which the animal moved;

model_move: A ModelMove instance;

t: An integer that defines the time step;

length: A float that defines the step length (i.e., the Euclidean distance between state_from (x, y) and state_to (x, y));

heading: A float that defines the angle (in polar coordinates) between state_from (x, y) and state_to (x, y);

Details

Patter.logpdf_step() is an internal generic function that evaluates the (unnormalised) log probability of an (unrestricted) movement step between two State(s) (i.e., locations). Methods are provided for the built-in State and ModelMove sub-types, but need to be provided for custom sub-types. Internally, Patter.logpdf_step() is wrapped by Patter.logpdf_move(), which evaluates the log probability of movement between two States, accounting for restrictions to movement; that is, Patter.logpdf_move() evaluates logpdf_step(state_from, state_to, model_move, length, heading) + log(abs(determinate)) - log(Z) where Z is the normalisation constant. This is required for particle smoothing (see particle_smoother_two_filter()).

Returns

A number (log probability);

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() and Patter.simulate_move() to simulate new States;

Patter.logpdf_step() and Patter.logpdf_move() to evaluate the log-probability of movement between two locations;

Patter.logpdf_move_normalisation() for estimation of the normalisation constant;

particle_smoother_two_filter() for the front-end function that uses these routines for particle smoothing;

Patter.particle_filter — Method

particle_filter(; timeline::Vector{DateTime},
                  xinit::Vector{<:State},
                  yobs::Dict,
                  model_move::ModelMove,
                  n_move::Int = 100_000,
                  n_record::Int = 1000,
                  n_resample::Float64 = Float64(n_record),
                  t_resample::Union{Nothing, Int, Vector{Int}}  = nothing,
                  n_iter::Int64 = 1,
                  direction::String = "forward", 
                  batch::Union{Nothing, Vector{String}} = nothing, 
                  progress = (), 
                  verbose::Bool = true)

A particle filtering algorithm that samples from f(s_t | y_{1:t}) for t ∈ 1:t.

Arguments (keywords)

timeline: A Vector{DateTime} of ordered, regularly spaced time stamps that defines the time steps for the simulation;

xinit: A Vector of State instances that defines the initial state(s) of the animal;

yobs: A Dictionary of observations:

Dictionary keys should match elements in timeline;

Each element must be a Vector of Tuples for that time step (one for each observation/sensor);

Each Tuple should contain (a) the observation and (b) the model parameters (that is, a ModelObs instance);

model_move: A ModelMove instance:

The movement model describes movement from one time step to the next and therefore depends implicitly on the resolution of timeline;

The movement model should align with the State instances in xinit. For example, a two-dimensional state (StateXY) requires a corresponding movement model instance (i.e., ModelMoveXY);

n_move: An integer that defines the number of attempts used to find a legal move;

All ModelMove sub-types contain a map field that defines the region(s) within which movements are allowed (see is_valid());

Each particle is moved up to n_move times, until a valid movement is simulated;

Particles that fail to generate a valid move are killed;

n_move = 1 is permitted, which corresponds to an 'unrestricted' movement model in which particles are instantly killed if they land in inhospitable habitats on map;

n_record: An integer that defines the number of particles to record at each time step:

n_record particles are resampled at each time step and recorded in memory;

n_resample: A number that defines the effective sample size for resampling:

Particles are resampled when the effective sample size <= n_resample;

t_resample: nothing, an integer or a Vector of integers that define the time step(s) at which to force resampling;

Particles are resampled at t_resample regardless of the effective sample size;

n_iter: A integer that defines the maximum number of iterations (trials);

direction: A String that defines the direction of the filter:

"forward" runs the filter forwards in time;

"backward" runs the filter backwards in time;

(optional) batch: A Vector of .jld2 file paths for particles (see Memory Management);

(optional) progress: A NamedTuple of arguments, passed to ProgressMeter.Progress, to control the progress bar. If enabled, one progress bar is shown for each batch;

verbose: A Bool that defines whether or not to print information to the console;

Algorithm

Initiation

The algorithm is initiated using a Vector of n_particle States (xinit). See simulate_states_init() to simulate initial states for the filter.

Movement

For every time step in the timeline, the internal function Patter.simulate_move() simulates the movement of particles away from previous States into new States using the movement model, as specified by model_move. Patter.simulate_move() is an iterative wrapper for a Patter.simulate_step() method that simulates a new State instance from the previous State. Patter.simulate_move() implements Patter.simulate_step() iteratively until a legal move is found (or n_move is reached). For custom State or ModelObs sub-types, a corresponding Patter.simulate_step() method is required. Illegal moves are those that land in NaN locations on the map or, in the case of States that include a depth (z) component, are below the depth of the seabed (see is_valid()). Particles that fail to generate legal moves are eventually killed by re-sampling (see below).

Likelihood

Observations are used to weight simulated particles. To simulate observations for filtering, use simulate_yobs(). To assemble real-world observations for filtering, see assemble_yobs(). For each valid State and time stamp in yobs, the log-probability of each observation, given the State, is evaluated via Patter.logpdf_obs(). For custom State or ModelObs sub-types, a corresponding Patter.logpdf_obs() method is required. The maximum weight across all particles (maxlp) is recorded at each time step as an algorithm diagnostic. (This metric can be intepreted as the maximum log-posterior if resampling is implemented at every time step.)

Resampling

Particles are periodically re-sampled, with replacement, using the low-variance systematic re-sampling algorithm (via Patter.resample()), at time steps in t_resample or when the effective sample size is less than or equal to n_resample. This has the effect of eliminating impossible particles and duplicating likely ones.

The algorithm continues in this way, iterating over the timeline, simulating, weighting and (re)sampling particles. At each time step, n_record particles are saved in memory. If the function fails to converge, a warning is returned alongside the outputs up to that time step. Otherwise, the function will continue to the end of the time series.

Multi-threading

The iteration over particles (i.e., simulated movements and likelihood evaluations) are multi-threaded.

Memory management

By default, n_record particles at each time step are retained in memory. If batch is provided, the timeline is split into length(batch) batches. The filter still moves along the whole timeline, but only records the particles for the current batch in memory. At the end of each batch, the particles for that batch are written to file. This reduces total memory demand.

batch file paths are sorted alphanumerically if direction = "forward" and in reverse order if direction = "backward". For example:

If you have a timeline of 10 time steps, direction = "forward" and batch = ["fwd-1.jld2", "fwd-2.jld2", "fwd-3.jld2"], fwd-1.jld2, fwd-2.jld2 and fwd-3.jld2 contain the particle matrices for time steps [1, 2, 3], [4, 5, 6] and [7, 8, 9, 10], respectively.

If you have a timeline of 10 time steps, direction = "backward" and batch = ["bwd-1.jld2", "bwd-2.jld2", "bwd-3.jld2"], bwd-1.jld2, bwd-2.jld2 and bwd-3.jld2 similarly contain the particle matrices for time steps [1, 2, 3], [4, 5, 6] and [7, 8, 9, 10], respectively.

Convergence and diagnostics

Algorithm convergence is not guaranteed. The algorithm may reach a dead-end–-a time step at which there are no valid locations into which the algorithm can step. This may be due to data errors, incorrect assumptions, insufficient sampling effort or poor tuning-parameter settings.

Returns

A Particles structure;

See also

State, ModelMove and ModelObs for State, movement model and observation model sub-types;

simulate_yobs() and assemble_yobs() to prepare observations for the particle filter;

Patter.simulate_step() and Patter.simulate_move() for the internal routines used to simulate new States;

Patter.logpdf_obs() methods to evaluate the log probability of observations;

particle_smoother_two_filter() to implement particle smoothing;

Patter.particle_smoother_two_filter — Method

particle_smoother_two_filter(; timeline::Vector{DateTime}, 
                               xfwd::Union{Matrix{<:State}, Vector{String}}, 
                               xbwd::Union{Matrix{<:State}, Vector{String}}, 
                               model_move::ModelMove, 
                               vmap::Union{GeoArray, Nothing} = nothing, 
                               n_sim::Int = 100, 
                               n_particle::Union{Nothing, Int} = nothing,
                               cache::Bool = true, 
                               batch::Union{Nothing, Vector{String}} = nothing, 
                               progress = true, 
                               verbose::Bool = true)

A two-filter particle smoother that samples from f(s_t | y_{1:T}) for t ∈ 1:T.

Arguments (keywords)

timeline: A Vector{DateTime} of ordered, regularly spaced time stamps that defines the time steps for the simulation;

xfwd, xbwd: Particles from the forward and backward filters (see particle_filter()), supplied as:

A Matrix of States (in memory);

A Vector of file paths, if particle_filter() was implemented with batch;

model_move: A ModelMove instance;

vmap: (optional) A GeoArray that defines the 'validity map' (see Patter.logpdf_move());

n_sim: An integer that defines the number of Monte Carlo simulations (see Patter.logpdf_move());

(optional) n_particle: An integer that defines the number of particles for smoothing. If nothing, all particles are used;

cache: A Bool that defines whether or not to precompute and cache movement density normalisation constants (see Patter.logpdf_move());

(optional) batch: A Vector of .jld2 file paths for particles (see particle_filter());

(optional) progress: A NamedTuple of arguments, passed to ProgressMeter.Progress, to control the progress bar. If enabled, one progress bar is shown for each batch;

verbose: A Bool that defines whether or not to print information to the console;

Details

particle_smoother_two_filter() smooths particles from the particle filter (see particle_filter()). The timeline from the particle filter should be supplied as well as a Matrix of particles from a forward run and a backward run (or a Vector of file paths to those matrices). The two filter smoother works by iteratively resampling particles in line with the probability density of movement between particles from the backward filter at time t and particles from the forward filter at time t - 1. Patter.logpdf_move() is an internal function that evaluates the log probability of a movement step between particles. This function wraps the Patter.logpdf_step() generic. Methods are provided for built-in State and ModelMove sub-types. To use custom sub-types, a corresponding Patter.logpdf_step() method should be provided. In particle_smoother_two_filter(), the vmap and n_sim arguments support the calculate of probability densities (see Patter.logpdf_move()). For movement models for which the density only depends on fields in xbwd and xfwd, set cache = true to precompute and store normalisation constants for density calculations for unique xbwd elements. Note that since typically only a subsample of particles from particle_filter() are retained in memory, it is not guaranteed that valid moves will exist between particle pairs at all time steps. At time step(s) in which the two filters are incompatible, 50 % of particles are retained from the forward filter and 50 % from the backward filter with a warning. The effective sample size at such time steps is set to NaN, providing an index and counter for problematic time steps (see Returns). Batching is only implemented if the inputs (xfwd, and xbwd) and outputs (via batch) are batched (and contain the same number of batches).

Returns

A Particles structure;

See also

particle_filter() implements the particle filter;

Patter.logpdf_step(), logpdf_move_normalisation() and Patter.logpdf_move() evaluate the log probability (density) of movement between two States;

particle_smoother_two_filter() implements the two-filter particle smoother;

Source

Fearnhead, P., Wyncoll, D., Tawn, J., 2010. A sequential smoothing algorithm with linear computational cost. Biometrika 97, 447–464.

Patter.resample — Function

resample(w::Vector{Float64}, n::Int)

(Internal) Given a weight vector w, resample a set of indices based on low-variance resampling algorithm from Thrun, Burgard and Fox's "Probabilistic Robotics".

Arguments

w: A Vector{Float64} of weights;

n: An integer that defines the number of particles;

Details

This is an internal function that implements systematic resampling in the particle filter (see particle_filter()) and smoothing algorithms (see particle_smoother_two_filter()). Note that for large n, the function is not numerically stable.

Returns

An integer vector of indices;

Example

X = ["A", "B", "C", "D"]
w = [0, 0, 0.75, 0.25]

idx = resample(w, 12)
X[idx]

Source

Code adapted from https://github.com/JuliaStats/StatsBase.jl/issues/124.

Patter.simulate_move — Function

simulate_move(state::State, model_move::ModelMove, t::Int64, n_trial::Real)

(Internal) Simulate movement from one location (State) into a new location (State).

Arguments

state: A State instance that defines the animal's previous State;

model_move: A ModelMove instance;

t: An integer that defines the time step;

n_trial: A number that defines the number of attempts to simulate a valid State;

Details

Patter.simulate_move() is an internal function that uses a Patter.simulate_step() method to simulate new state proposals iteratively until a valid State is generated or n_trial is reached (see is_valid()). For custom States or ModelMove sub-types, corresponding Patter.simulate_step() methods must be provided for this function. Patter.simulate_move() is used to simulate movement paths (e.g., in simulate_path_walk() and particle_filter()).

Returns

A Tuple that comprises the simulated State instance and the (log) weight; i.e.,

([State], 0.0) if State is valid;

([State], -Inf) if State is invalid;

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() to simulate a new State;

Patter.is_valid() to determine whether or not a simulated state is valid;

Patter.simulate_move() to simulate states iteratively until a valid state is found;

simulate_path_walk() and particle_filter() for the front-end functions that use these routines to simulate animal movement paths;

Patter.simulate_path_walk — Method

simulate_path_walk(; xinit = Vector, model_move::ModelMove, timeline::Vector{DateTime})

Simulate discrete-time movement path(s) from a Vector of initial States and a random-walk movement model.

Arguments (keywords)

xinit: A Vector of initial States instances;

model_move: A ModelMove instance;

timeline: A Vector{DateTime} of ordered, regularly spaced time stamps that defines the time steps for the simulation;

Details

State refers to the (x, y) location of an animal (alongside additional state components, if applicable). To simulate initial states, use simulate_states_init(). For each initial state, simulate_path_walk() simulates a sequence of States (i.e., a movement path) of length(timeline) steps using the movement model (model_move). The simulation of movement from one State into another is implemented by the internal function Patter.simulate_move(), which in turn wraps Patter.simulate_step(). At each time step, Patter.simulate_move() implements Patter.simulate_step() iteratively until a valid movement is identified (see is_valid()). Patter.simulate_step() is a generic function. Methods are implemented for the built-in State and ModelMove sub-types but custom sub-types require a corresponding Patter.simulate_step() method.

Returns

A matrix of States:

Each row represents a simulated path;

Each column represents a time step along timeline;

See also

Patter.simulate_states_init — Function

simulate_states_init(; map::Rasters.Raster,
                       timeline::Vector{DateTime}, 
                       state_type::Type{<: State}, 
                       xinit::Union{Nothing, DataFrame}, 
                       model_move::ModelMove, 
                       datasets::Union{Nothing, Vector}, 
                       model_obs_types::Union{Nothing, Vector}, 
                       n_particle::Int, 
                       direction::String = "forward", 
                       output = "DataFrame")

Simulate a DataFrame or Vector of initial States for the simulation of movement paths in simulate_path_walk() and particle_filter().

Arguments

map: A Rasters.Raster that defines the study area for the simulation. Here, map is used to:

Sample initial coordinates, via Patter.coords_init(), if xinit = nothing;

timeline, model_move, datasets, model_obs_types, direction: Additional arguments used to restrict map, via Patter.map_init(), before sampling initial states.

timeline: A sorted, DateTime vector of regularly spaced time stamps that defines the timeline for the simulation;

model_move: A ModelMove instance;

(optional) datasets: A Vector of observation datasets or nothing;

(optional) model_obs_types: A Vector of ModelObs sub-types or nothing;

direction: A String string that defines the direction of the simulation ("forward" or "backward");

state_type: The State sub-type. Here, state is used to:

Convert sampled coordinates to initial states, via Patter.states_init(), if xinit = nothing;

xinit: (optional) A DataFrame of initial states, with one column for each state dimension.

n_particle: An integer that defines the number of simulated states:

If xinit = nothing, n_particle specifies the number of simulated states via Patter.coords_init();

If xinit is supplied but there are not n_particle initial states, n_particle initial states are re-sampled from xinit with replacement;

output: A String that defines the output format:

"DataFrame" returns a DataFrame (for communication with R);

"Vector" returns a Vector of `State(s);

Details

These functions support the simulation of initial states for animal movement walks in simulate_path_walk() and particle_filter().

If xinit = nothing, initial coordinates are sampled from map.

The region(s) within map from which initial coordinates are sampled can be optionally restricted by the provision of the observation datasets and the associated model sub-types (via Patter.map_init_iter()). This option does not apply to simulate_path_walk() but is used in particle_filter() where observation models are used. In this instance, Patter.map_init_iter() iterates over each model and uses the Patter.map_init() method to update map. The following methods are implemented:

Default. The default method returns map unchanged.

model_obs_type::ModelObsAcousticLogisTrunc. This method uses acoustic observations to restrict map via Lavender et al.'s (2023) acoustic–container algorithm. The function identifies the receiver(s) that recorded detection(s) immediately before, at and following the first time step (timeline[start], where start is 1 if direction = "forward" and length(timeline) otherwise). The 'container' within which the individual must be located from the perspective of each receiver is defined by the time difference and the individual's mobility (that is, the maximum moveable distance the individual could move between two time steps), which must be specified in model_move.mobility. The intersection between all containers defines the possible locations of the individual at the first time step.

model_obs_type::ModelObsDepthUniformSeabed. This method uses the depth observations to restrict map (which should represent the bathymetry in a region). The individual must be within a region in which the observed depth at timeline[start] is within a depth envelope around the bathymetric depth defined by the parameters depth_shallow_eps and depth_deep_eps (see ModelObs). (If there is no observation at timeline[start], map is returned unchanged.)

model_obs_type::ModelObsDepthNormalTruncSeabed. This method also uses depth observations to restrict map. The individual must be in a location where the bathymetric depth plus the depth_deep_eps parameter at timeline[start] is greater than or equal to the observed depth at timeline[start] (see ModelObs). (If there is no observation at timeline[start], map is returned unchanged.)

To handle custom ModelObs sub-types, process map beforehand or write an appropriate Patter.map_init() method.

Using map, a DataFrame of n_particle initial coordinates (map_value, x, y) is sampled using Patter.coords_init(). Additional state dimensions are added, as required depending on the state_type, via a Patter.states_init() method. For custom State sub-types, a corresponding Patter.states_init() method is required (or supply xinit yourself).

If xinit() is provided and n_particle initial states are provided, xinit is returned unchanged. Otherwise, n_particle initial states are resampled from xinit, with replacement, and returned.

Returns

simulate_states_init() returns:

If output = "DataFrame", a DataFrame, with n_particle rows, and one column for each State dimension;

If output = "Vector", a Vector of State instances, with n_particle elements;

See also

These functions are used to initialise simulated movement trajectories in simulate_path_walk() and particle_filter(). See also:

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() and Patter.simulate_move() to simulate new States;

simulate_path_walk() to simulate animal movement paths (via ModelMove);

simulate_yobs() to simulate observations arising from simulated movements (via ModelObs);

Patter.simulate_step — Function

simulate_step(state::State, model_move::ModelMove, t::Int64)

(Internal) Simulate a (tentative) step from one location (State) into a new location (State).

Arguments

state: A State instance that defines the animal's previous State;

model_move: A ModelMove instance;

t: An integer that defines the time step;

Details

Patter.simulate_step() is an internal generic function that simulates a new value for the animal's [State], that is, the animal's location (and other state components). Methods are provided for the built-in [State] and movement model (ModelMove) sub-types. For custom States or ModelMove sub-types, corresponding methods must be provided. Internally, Patter.simulate_step() is wrapped by Patter.simulate_move(), which implements simulate_step() iteratively until a valid State is simulated (see is_valid()).

Returns

A State instance;

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() to simulate a new State;

Patter.is_valid() to determine whether or not a simulated state is valid;

Patter.simulate_move() to simulate states iteratively until a valid state is found;

simulate_path_walk() and particle_filter() for the front-end functions that use these routines to simulate animal movement paths;

Patter.simulate_yobs — Method

simulate_yobs(; paths::Matrix{<:State}, model_obs::Vector{ModelObs}, timeline::Vector{DateTime})

For a series of simulated paths, simulate a dictionary of observations.

Arguments

paths: A Matrix of simulated paths from simulate_path_walk();

model_obs: A Vector of ModelObs instances;

timeline: A Vector{DateTime} of ordered, regularly spaced time stamps that defines the time steps for the simulation;

Details

The function expects a Matrix of simulated paths (see simulate_path_walk()). For each simulated path, the function iterates over each step in timeline and simulates observations using the Vector of observation models. Observations are simulated by the internal generic Patter.simulate_obs() via simulate_obs(State, model, t), where t is the time step. Methods are provided for the built-in States and ModelObs sub-types. For custom sub-types, a corresponding Patter.simulate_obs() method is required. Simulated observations can be used in the particle filter to reconstruct the underlying movements (see particle_filter()).

Returns

A Dict, with one entry for each path:

Each entry is a Dict, with one entry for each time stamp;

Each time stamp entry is a Vector of Tuples, each comprising the simulated observation and the associated ModelObs instance (see also assemble_yobs());

See also

State and ModelMove for State and movement model sub-types;

Patter.simulate_step() and Patter.simulate_move() to simulate new States;

simulate_path_walk() to simulate animal movement paths (via ModelMove);

simulate_yobs() to simulate observations arising from simulated movements (via ModelObs);

Patter.ModelMove — Type

Movement models

ModelMove is an Abstract Type that groups movement models.

Built-in sub-types

ModelMove sub-types define the components of different kinds of movement model. The following sub-types are built-in:

Random walks (RWs): - ModelMoveXY(map, mobility, dbn_length, dbn_heading): A sub-type for two-dimensional (x, y) random walks, based distributions for step lengths (dbn_length) and headings (dbn_heading); - ModelMoveXYZ(map, mobility, dbn_length, dbn_heading, dbn_z): A subtype for a RW in X, Y and Z, based on distributions for step lengths (dbn_length), headings (dbn_heading) and depth (dbn_z);

Correlated random walks (CRWs): - ModelMoveCXY(map, mobility, dbn_length, dbn_heading_delta): A sub-type for a CRW in X and Y, based on distributions for step lengths (dbn_length) and turning angles (dbn_heading_delta) ; - ModelMoveCXYZ(map, mobility, dbn_length, dbn_heading_delta, dbn_z_delta): A sub-type for a CRW in X, Y and Z, based on distributions for step lengths (dbn_length), turning angles (dbn_heading_delta) and changes in depth (dbn_z_delta);

These contain the following fields:

map: A field that defines the arena within which movement occurs. The coordinate reference system of the map must align with the other components of the movement model, which typically require a Universal Transverse Mercator (planar) projection with coordinates in metres. map is required by all movement models;

mobility: A number that defines the maximum movement distance between two time steps;

dbn_length: The distribution of step lengths;

dbn_heading: The distribution of headings;

dbn_heading_delta: The distribution of turning angles;

dbn_z: The distribution of depths;

dbn_z_delta: The distribution of changes in depth;

Custom sub-types

To define a custom sub-type, such as ModelMoveXYZ*, simply define a struct that is a sub-type of Patter.ModelMove:

struct ModelMoveXYZ{T, U, V, W, X} <: Patter.ModelMove
    # The environment (i.e., map)
    # > This defines the regions within which movements are permitted (i.e., in water)
    map::T
    # Distribution for step lengths
    mobility::U
    dbn_length::V
    # Distribution for headings
    dbn_heading::W
    # Distribution for changes in depth
    dbn_z::X
  end

Note this structure is now implemented.

New ModelMove structures should obey the following requirements:

The map and mobility fields are required by all ModelMove sub-types;

By default, map is assumed to be a GeoArray but a shapefile can be used with a custom extract() method;

To use a new ModelMove sub-type in the simulation of animal movements (via simulate_path_walk()) and particle-filtering algorithms, the following steps are also necessary:

Define a corresponding State sub-type;

(optional) Define Patter.map_init() and Patter.states_init() methods for simulate_states_init() to simulate initial states;

Define a Patter.simulate_step() method (for Patter.simulate_move()) to update the state using a ModelMove instance (in simulate_path_walk() and particle_filter());

Define a Patter.logpdf_step() method (for Patter.logpdf_move()) to evaluate the probability density of movement from one state to another (in particle_smoother_two_filter());

Patter.ModelObs — Type

Observation models

ModelObs

ModelObs is an Abstract Type that groups observation model structures. See below for built-in sub-types.

Built-in sub-types

ModelObsAcousticLogisTrunc

ModelObsAcousticLogisTrunc is a ModelObs structure for an acoustic observation (0, 1) and a truncated logistic detection probability model. This contains the following fields:

sensor_id: An integer that defines the sensor (receiver) ID;

receiver_x, receiver_y: Floats that define the x and y coordinates of the receiver;

receiver_alpha, receiver_beta, receiver_gamma: Floats that define the parameters of a truncated logistic detection probability model.

An acoustic observation ($y^{(A)}_{t, k} \in {0, 1}$) at receiver $k$ (location $\textit{\textbf{r}}_k = (\text{receiver\_x}, \text{receiver\_y})$) at time $t$ is modelled using a Bernoulli probability mass function:

\[f(y^{(A)}_{t, k} | \textit{\textbf{s}}_t) = \text{Bernoulli}(p_{k,t}(\textit{\textbf{s}}_t))\]

where $p_{k,t}(\textit{\textbf{s}}_t)$ is the probability of a detection at receiver $k$ at time $t$ given a transmission from location $\textit{\textbf{s}}_t = (x, y)$. ModelObsAcousticLogisTrunc represents $p_{k,t}(\textit{\textbf{s}}_t)$ as a logistic function of the Euclidean location between the receiver and the transmitter, according to the equation:

\[p_{k,t}(\textit{\textbf{s}}_t) = \left\{ \begin{array}{ll} (1 + e^{-(\text{receiver\_alpha} - \text{receiver\_beta} \cdot |\textit{\textbf{s}}_t - \textit{\textbf{r}}_k|)})^{-1} & \text{if } |\textit{\textbf{s}}_t - \textit{\textbf{r}}_k| < \text{receiver\_gamma} \\ 0 & \text{otherwise} \end{array} \right.\]

where $\text{receiver\_gamma}$ is the detection range.

To simulate an acoustic observation ($y^{(A)}_{t, k} \in {0, 1}$) from this model, we can draw a sample from a Bernoulli distribution:

\[y^{(A)}_{t, k} | \textit{\textbf{s}}_t \sim \text{Bernoulli}(p_{k,t}(\textit{\textbf{s}}_t))\]

via Patter.simulate_obs().

ModelObsDepthUniformSeabed

ModelObsDepthUniformSeabed is ModelObs structure for a depth observation and a uniform depth model. This contains the following fields:

sensor_id: An integer that defines the sensor (tag) ID;

depth_shallow_eps: A float that defines the shallow depth error;

depth_deep_eps: A float that defines the deep depth error;

This model assumes that an individual must be located in an envelope around the seabed, defined by two error terms (depth_shallow_eps and depth_deep_eps), according to the equation:

\[f\left( y_t^{(D)} | \textit{\textbf{s}}_t \right) = \begin{cases} z_t & \text{if } b(\textit{\textbf{s}}_t) - \text{depth\_shallow\_eps} \leq y_t^{(D)} \leq b(\textit{\textbf{s}}_t) + \text{depth\_deep\_eps} \\ 0 & \text{otherwise} \end{cases}\]

where $y_t^{(D)}$ is the observed depth, $b(\textit{\textbf{s}}_t)$ is the bathymetric depth in location $\textit{\textbf{s}}_t$ and $z_t$ is a constant. If depth_shallow_eps and depth_deep_eps are zero, the individual's depth is required to match the bathymetric depth.

We can simulate observations from this model as follows:

\[y_t^{(D)} | \textit{\textbf{s}}_t \sim \text{Uniform}(b(\textit{\textbf{s}}_t) + \text{depth\_deep\_eps}, \text{min}(b(\textit{\textbf{s}}_t) - \text{depth\_shallow\_eps}, 0))\]

via Patter.simulate_obs(). If depth_shallow_eps and depth_deep_eps are set to zero, the Patter.simulate_obs() method simply returns the bathymetric depth (state.map_value).

ModelObsDepthNormalTruncSeabed

ModelObsDepthNormalTruncSeabed is a ModelObs structure for a depth observation and a truncated normal model. This contains the following fields:

sensor_id: An integer that defines the sensor (tag) ID;

depth_sigma: A float that defines the standard deviation of the normal distribution;

depth_deep_eps: A float that defines the deep truncation parameter;

This model assumes that an individual must be located in an envelope around the seabed, defined by a normal distribution centred at this location with standard deviation depth_sigma:

\[f(y_t^{(D)} | \textit{\textbf{s}}_t) = \text{TruncatedNormal}(b(\textit{\textbf{s}}_t), \text{depth\_sigma}^2, 0, b(\textit{\textbf{s}}_t)).\]

We can simulate observations from this model as for previous models via Patter.simulate_obs().

ModelObsContainer

ModelObsContainer is a ModelObs structure for a container. Containers are a computational device used to mitigate particle degeneracy in the particle filter. Containers define the maximum possible distance of an individual from a location (e.g., receiver) that recorded a future observation. Accordingly, the ModelObsContainer structure contains the following fields:

sensor_id: An integer that defines the sensor ID (e.g., the receiver ID for the receiver that recorded the next detection);

centroid_x, centroid_y: Floats that define the x and y coordinates of the container's centroid;

radius: A Float that defines the radius of the container;

An acoustic container defines the region within which an individual must be located according to the receiver(s) at which it was next detected. In the particle filter, particles are permitted or killed depending on whether or not they are compatible with this constraint.

We formalise the acoustic container constraint as follows. Let $t$ denote the movement timeline (where $t \in \{1,2, \dots ,T\}$), $\boldsymbol{s}$ denote the individual's latent state (typically (x,y) location), $\boldsymbol{y}$ denote acoustic observations (detections, non-detections) at receivers and $\boldsymbol{\theta}$ denote static parameters (in the movement and observation models). The likelihood of the observations $\boldsymbol{y}_{1:T}$ given the parameters $\boldsymbol{\theta}$ is expressed as

\[f(\boldsymbol{y}_{1:T}\mid\boldsymbol{\theta}) = \int \prod_{t=1}^{T} f(\boldsymbol{y}_{t}\mid\boldsymbol{s}_{t},\boldsymbol{\theta}) \, f(\boldsymbol{s}_{t}\mid\boldsymbol{s}_{t-1},\boldsymbol{\theta}) \, d\boldsymbol{s}_{1:T}.\]

Acoustic containerisation represents a redefinition of the likelihood $f(\boldsymbol{y}_{t}\mid\boldsymbol{s}_{t},\boldsymbol{\theta})$, in a manner that does not affect the likelihood $f(\boldsymbol{y}_{1:T}\mid\boldsymbol{\theta})$, with the addition of a characteristic function $\chi(t,\boldsymbol{s}_{t},\boldsymbol{y}) \in \{0,1\}$ that returns zero or one depending on whether or not the latent state $\boldsymbol{s}_{t}$ is compatible with the observations at some future time:

\[f(\boldsymbol{y}_{1:T}\mid\boldsymbol{\theta}) = \int \prod_{t=1}^{T} f(\boldsymbol{y}_{t}\mid\boldsymbol{s}_{t},\boldsymbol{\theta}) \, \chi(t,\boldsymbol{s}_{t},\boldsymbol{y}) \, f(\boldsymbol{s}_{t}\mid\boldsymbol{s}_{t-1},\boldsymbol{\theta}) \, d\boldsymbol{s}_{1:T}.\]

The characteristic function $\chi(t,\boldsymbol{s}_{t},\boldsymbol{y})$ is defined as follows. Let $\Delta(t)$ be a function that identifies the next time step (at some time after $t$) at which detection(s) were recorded. At time $t$ the individual must be within a maximum distance (denoted $\text{radius}_{k}$) of each receiver $k$ that recorded the next detection(s):

\[\chi(t,\boldsymbol{s}_{t},\boldsymbol{y}) = \prod_{k} \chi_{k}\!\bigl( \lVert \boldsymbol{r}_{k,\Delta(t)} - \boldsymbol{s}_{t} \rVert \le \text{radius}_{k}(\text{mobility}) \bigr),\]

where $\boldsymbol{r}_{k,\Delta(t)}$ denotes the location of receiver $k$.

The radius for receiver $k$ at time $t$ depends on the number of time steps from $t$ until the detection event $(\Delta(t)-t)$, a parameter $\text{mobility} \in \boldsymbol{\theta}$ that defines the maximum possible moveable distance in between two time steps (i.e., from $t \to t+1$) and the receiver’s maximum detection range $\gamma \in \boldsymbol{\theta}$:

\[\text{radius}(\text{mobility}) = (\Delta(t)-t)\,\text{mobility} + \gamma.\]

In practice, acoustic containers are implemented via the ModelObsContainer structure (that is, we treat the container like an additional observation, even though it is simply a redundant use of the data). ModelObsContainer instances must be assembled before the filter run, following the equations above. For relevant time steps, it is necessary to build a ModelObsContainer for each of the receiver(s) that recorded a detection at the next detection time step. Each instance contains centroid_x and centroid_y elements that define the centroid of the container and a radius element that defines the maximum possible distance of the individual from the container centroid (see above).

In the particle filter, at each time $t$, we iterate over all observation‑model structures and update particle weights using a Patter.logpdf_obs method. For ModelObsContainer instances, the Patter.logpdf_obs method computes the distance between the particle $(i)$ and receiver $k$’s location (centroid) and adds $0.0$ or $\text{-Inf}$ to the log weight ($\text{lw}$) depending on whether or not the distance is less than the radius:

\[\text{lw}_i = \text{lw}_i + \log\!\bigl( \chi_{k}\!\bigl( \lVert \boldsymbol{r}_{k,\Delta(t)} - \boldsymbol{s}_{t} \rVert \le \text{radius}_{k}(\text{mobility}) \bigr) \bigr).\]

This approach is valid because we only kill particles that are incompatible with future observations. A filtering–smoothing algorithm that implements acoustic containerisation thus still approximates $f(\boldsymbol{s}_{t}\mid\boldsymbol{y}_{1:T})$ once all data have been taken into account.

Custom sub-types

To define a custom sub-type, such as ModelObsDepthNormal, simply define a struct that is a sub-type of Patter.ModelObs:

struct ModelObsDepthNormal <: Patter.ModelObs
    sensor_id::Int64
    depth_sigma::Float64
end

For communication with R, all sub-types should include a sensor_id field.

Add corresponding methods to simulate observations via Patter.simulate_obs() and to evaluate log probabilities via Patter.logpdf_obs().

Simulation

Patter.simulate_obs() is an internal generic function that simulates observations, given the animal's State and a ModelObs instance. This accepts the following arguments:

state: A State instance;

model_obs: A ModelObs instance;

t: An integer that defines the time step;

Methods are implemented for all built-in sub-types. Methods can be defined for new sub-types, such as ModelObsDepthNormal, as follows:

function Patter.simulate_obs(state::StateCXYZ, model_obs::ModelObsDepthNormal, t::Int64)
    dbn   = truncated(Normal(state.z, model_obs.depth_sigma), 0, state.map_value)
    rand(dbn)
end

Patter.simulate_obs() is wrapped by simulate_yobs() for the simulation of observations.

Log probabilities

Patter.logpdf_obs() is a generic function that calculates the log probability (density) of an observation, given the animal's State and a ModelObs instance. This accepts the following arguments:

state: A State instance;

model_obs: A ModelObs instance;

t: An integer that defines the time step;

obs: The observation;

Methods are implemented for all built-in sub-types. Methods can be defined for new sub-types, such as ModelObsDepthNormal, as follows:

function Patter.logpdf_obs(state::State, model_obs::ModelObsDepthNormal, t::Int64, obs::Float64)
    dbn   = truncated(Normal(state.map_value, model_obs.depth_sigma),
                      0.0, state.map_value)
    logpdf(dbn, obs)
  end

Patter.logpdf_obs() is used in particle_filter() to evaluate the log-probability of the data given particle samples.

Patter.Particles — Type

Particles(states::Union{Nothing, Matrix{<:State}}, diagnostics::DataFrame, callstats::DataFrame)

Fields

states: (optional) A Matrix of States:

Each row corresponds to a particle;

Each column corresponds to the timestep;

diagnostics: A DataFrame of algorithm diagnostics:

timestep: A Vector{Int64} of time steps;

timestamp: A Vector{DateTime} of time stamps;

ess: A Vector{Float64} that defines the effective sample size at each time step;

maxlp: Vector{Float64} that defines the maximum log weight at each time step (i.e., the maximum log-posterior, if resampling is implemented at every time step);

callstats: A one-row DataFrame of call statistics:

timestamp: A DateTime that define the start time of the function call;

routine: A String that defines the algorithm;

n_particle: An Int that defines the number of particles;

n_iter: An Int or NaN that defines the number of iterations (trials);

convergence: A Boolian that defines convergence;

loglik: A Float64 that defines the log-likelihood (log P(observations | θ));

time: A Float64 that defines the duration (s) of the function call;

Details

states is nothing if particle_filter() or particle_smoother_two_filter() are implemented with batching.

convergence is defined as follows:

In particle_filter(), convergence defines whether or not the filter reached the end of the timeline;

In particle_smoother_two_filter(), convergence defines whether or not correct smoothing was achieved on at least 95 % of time steps. 'Correct smoothing' is possible when there at at least some valid moves between the subset of recorded particles on the backward filter and those on the forward filter (for the previous time step).

See also

Particles objects are returned by:

particle_filter()

particle_smoother_two_filter()

Patter.State — Type

State

State is an Abstract Type that defines the animal's state at a given time step.

Built-in sub-types

The following sub-types are built-in:

StateXY(map_value, x, y): Used for two dimensional (x, y) states;

StateXYZ(map_value, x, y, z): Used for three-dimensional (x, y, z) states;

StateCXY(map_value, x, y, heading): Used for two-dimensional (x, y) correlated states;

StateCXYZ(map_value, x, y, z, heading): Used for three-dimensional (x, y, z) correlated states;

These contain the following fields:

map_value: The value of the movement map at coordinates (x, y), required for all States (see ModelMove);

x, y: Floats that define the animal's x and y coordinates, required for all States;

z: A float that defines the animal's z (depth) coordinate, required for all States with a depth component;

heading: A float that defines the heading, required by StateCXYZ;

Custom sub-types

To define a custom sub-type, such as StateXYZ*, simply define a struct that is a sub-type of Patter.State:

struct StateXYZ <: Patter.State
    # Map value
    map_value::Float64
    # Coordinates
    x::Float64
    y::Float64
    z::Float64
end

Note that StateXYZ is now implemented.

New states should obey the following requirements:

All states must include map_value, x and y fields;

For states with a depth dimension, the depth field must be named z (for Patter.simulate_move());

For R users, all fields must be of type Float64 for Patter.r_get_states() to parse state vectors;

To use a new State sub-type in the simulation of animal movements (via simulate_path_walk()) and particle-filtering algorithms, the following steps are also necessary:

Define a corresponding ModelMove sub-type;

(optional) Define Patter.map_init() and Patter.states_init() methods for simulate_states_init() to simulate initial states;

Define a Patter.simulate_step() method (for Patter.simulate_move()) to update the state using a ModelMove instance (in simulate_path_walk() and particle_filter());

Define a Patter.logpdf_step() method (for Patter.logpdf_move()) to evaluate the probability density of movement from one state to another (in particle_smoother_two_filter());

Index

Patter.ModelMove

Patter.ModelObs

Patter.Particles

Patter.State

Patter.assemble_yobs

Patter.extract

Patter.is_valid

Patter.logpdf_move

Patter.logpdf_move_normalisation

Patter.logpdf_step

Patter.particle_filter

Patter.particle_smoother_two_filter

Patter.resample

Patter.simulate_move

Patter.simulate_path_walk

Patter.simulate_states_init

Patter.simulate_step

Patter.simulate_yobs